+++ test/basic-loader-test/index.js
@@ -1,0 +1,4 @@
+/* eslint-disable import/no-unresolved */
+// some file
+// eslint-disable-next-line import/extensions
+import './file.js?q=1#hash';