+++ test/basic-loader-test/mod2.js
@@ -1,0 +1,3 @@
+const test = require('./mod1');
+
+module.exports = new URL('./style.less', import.meta.url);