+++ test/css-loader-example/index.js
@@ -1,0 +1,4 @@
+/* eslint-disable import/no-unresolved */
+// some file
+import './style.css';
+import './style.modules.css';